Added User controller goodies.

+++ src/main/java/org/lovepacs/repositories/UserRepository.java
@@ -3,9 +3,9 @@ package org.lovepacs.repositories;
 import org.lovepacs.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
-    User findById(long id);
+    User findById(Integer id);
 
     User findByName(String name);
 }
